ml-service/requirements.txt
# ML Service Dependencies for Fraud Detection

# Core ML Libraries
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Deep Learning (optional)
tensorflow==2.14.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Blockchain Integration
stacks-blockchain==0.1.0

# Data Processing
joblib==1.3.2
scipy==1.11.4

# Merkle Tree
pymerkle==4.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
